Reverting some uint32 changes

+++ Source/core/IPCMessage.h
@@ -258,11 +258,11 @@ namespace Core {
                 : _buffer()
             {
             }
-            inline BufferType(const uint32_t length)
+            inline BufferType(const uint16_t length)
                 : _buffer(length)
             {
             }
-            inline BufferType(const uint32_t length, const uint8_t buffer[])
+            inline BufferType(const uint16_t length, const uint8_t buffer[])
                 : _buffer(buffer, length)
             {
             }
@@ -282,17 +282,17 @@ namespace Core {
             {
                 return (&(_buffer[0]));
             }
-            inline void Set (const uint32_t length, const uint8_t buffer[]) {
+            inline void Set (const uint16_t length, const uint8_t buffer[]) {
                 _buffer.Copy(0, length, buffer);
             }
-            inline uint16_t Serialize(uint8_t buffer[], const uint32_t length, const uint32_t offset) const
+            inline uint16_t Serialize(uint8_t buffer[], const uint16_t length, const uint16_t offset) const
             {
                 uint16_t size = ((_buffer.Size() - offset) > length ? length : (_buffer.Size() - offset));
                 ::memcpy(buffer, &(_buffer[offset]), size);
 
                 return (size);
             }
-            uint16_t Deserialize(const uint8_t buffer[], const uint32_t length, const uint32_t offset)
+            uint16_t Deserialize(const uint8_t buffer[], const uint16_t length, const uint16_t offset)
             {
                 ASSERT (offset == _buffer.Size());
